Added file load button to have data to feed the lib
Added a MidiFileParser class. For now, it only opens a MidiFile (from the leff MidiFile lib).
The input to open the MidiFile is an InputStream obtained from a ContentResolver's reading of a URI obtained from an ActivityResult after launching the file open activity.
This logic is done in the MainActivity for now. Cleaner implementation would give it its own activity.

An empty NativeManager lib should handle the two points of communication between the app and the lib :

- Convert a NoteOnEvent or NoteOffEvent to a NoteEvent for the C++ lib while parsing and/or push that event to a renderer
- Call perform() from the renderer in the app and obtain the equivalent to an std::vector<Set<NoteEvent>>.

This is the heart of the issue we're approaching, and it has to do with how the NDK works. It's terra incognita from now on.