First attempt at using JavaCpp to interface with the native lib. It fails miserably.
The NativeManager class has been moved to a package meant to handle native interaction.
It shall be the only gateway for the rest of the program to access native functionality.

Instead of vanilla JNI, I attempted to make use of the JavaCPP library.
Its idea is to write the glue in Java instead of C++, and then compile Java code into JNI glue.

Therefore, two classes that mirror the lib's CommandData and NoteData structs were created.
The NativeManager now has a class to convert a MidiMessage into a CommandData. Theoretically, it would be used in the main loop when a command is detected.
The MidiMessage class has been expanded with the appropriate util functions.
And the MainActivity now calls on this function to convert the generated MidiMessage at the press of the main button.

Problem is : this compiles, but crashes immediately when the button is pressed.
This is because, as is, JavaCPP does not actually compile the Java code into JNI glue.
For that, additional compilation steps are necessary, and in the case of Gradle, that means a Gradle plugin (because native Gradle can't run Java compile commands, oopsie)

There is indeed a plugin to connect JavaCPP to Gradle, aptly named Gradle-JavaCPP.
However, it does not compile on my system : gradle -build at the root of the plugin repo simply fails.

Although I did add the necessary lines to the app's build.gradle to use Gradle-JavaCPP if it compiles (approximately : they are copy-pasted from the plugin's README and need to be adapted), they're commented out because the .jar for the plugin doesn't exist, since I can't compile it.

This is where it is. Thus, we need to :

1. Find a way to compile Gradle-JavaCPP
2. Configure Gradle with this plugin to properly apply JavaCPP.
3. Utilize JavaCPP to properly connect to the library.

This is the hard part. We're gonna be stuck here a while.