File loading button implemented (MPTK pro acquired)
Changed MPTK to pro version to permit arbitrary file loading. Added the button in the scene to load up the file. Added a library called NativeFilePicker to do just that ; it works on Windows and is normally designed to also work on Android.

Now, when the app starts it will not have a preloaded file. You have to find a file and load it in. Of course we can change it to default sometime later.