Fixed ticks that were absolute, and changed to milliseconds
I had somehow completely forgotten to convert the ticks to relative, which resulted in no chords being played. This is fixed.

After confirmation on the MPTK forum, I am now also using the RealTime field which is an absolute ms tick, so I don't have to do the conversion myself.  This is so that the proper temporal resolution is enabled on the C++ side to eliminate anti-shadows.

Also made it so that the sound stops when loading in a new file (obviously).