[MINOR] Use file presets, but other bugs remain
We now load the presets provided by the MIDI file. This is nothing but a workaround because it only uses the latest presets. It also doesn't apply the controller changes that go with them, because these are meta events...maybe later ? The good way to do this would be to have it in the C++ library.

Tried excluding UI buttons from click area. But it fails. The recommended way found online has the effect of canceling any detection of touches except for releases.

At any rate, touch controls are broken. The algorithm used instead of the commented version is broken. We should NOT have to do this. We should be able to just use the touch phase. There is something wrong with the fact that holding a finger to the screen sends press messages every frame and not stationary.